[lexical-playground] Refactor: Migrate PageBreakPlugin to PageBreakExtension
Converts the legacy React PageBreakPlugin into a declarative
PageBreakExtension using defineExtension, registering the node and
INSERT_PAGE_BREAK command through the extension system.
